+++ src/Migration/Gateway/HttpSimpleClient.php
@@ -41,7 +41,6 @@ class HttpSimpleClient implements HttpClientInterface
     protected function constructClient(array $additionalOptions = []): Client
     {
         return new Client(\array_merge([
-            'verify' => false,
             'connect_timeout' => 15.0,
         ], $additionalOptions));
     }