+++ src/Profile/Shopware/Gateway/Api/Reader/EnvironmentReader.php
@@ -8,11 +8,9 @@
 namespace SwagMigrationAssistant\Profile\Shopware\Gateway\Api\Reader;
 
 use GuzzleHttp\Exception\ClientException;
-use GuzzleHttp\Exception\ConnectException;
 use GuzzleHttp\Exception\RequestException as GuzzleRequestException;
 use Psr\Http\Message\ResponseInterface;
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Core\Framework\ShopwareHttpException;
 use SwagMigrationAssistant\Exception\MigrationException;
 use SwagMigrationAssistant\Migration\Gateway\HttpClientInterface;
 use SwagMigrationAssistant\Migration\Gateway\Reader\EnvironmentReaderInterface;
@@ -30,8 +28,6 @@ class EnvironmentReader implements EnvironmentReaderInterface
 {
     private ?HttpClientInterface $client = null;
 
-    private MigrationContextInterface $migrationContext;
-
     public function __construct(private readonly ConnectionFactoryInterface $connectionFactory)
     {
     }
@@ -41,7 +37,6 @@ class EnvironmentReader implements EnvironmentReaderInterface
      */
     public function read(MigrationContextInterface $migrationContext): array
     {
-        $this->migrationContext = $migrationContext;
         $this->client = $this->connectionFactory->createApiClient($migrationContext);
 
         $information = [
@@ -49,161 +44,83 @@ class EnvironmentReader implements EnvironmentReaderInterface
             'requestStatus' => new RequestStatusStruct(),
         ];
 
-        if ($this->client === null) {
-            $information['requestStatus'] = new RequestStatusStruct('SWAG-EMPTY-CREDENTIALS', 'Empty credentials');
-
-            return $information;
-        }
-
-        if ($this->doSecureCheck($information)) {
-            return $information;
-        }
-
-        $requestStatus = $information['requestStatus'];
-
-        if ($requestStatus->getCode() === MigrationException::sslRequired()->getErrorCode()) {
-            $requestStatus->setIsWarning(false);
-
-            return $information;
-        }
-
-        if ($this->doInsecureCheck($information)) {
-            return $information;
-        }
-
-        if ($this->doShopwareCheck($information)) {
-            return $information;
+        try {
+            $information['environmentInformation'] = $this->getEnvironmentInformation();
+        } catch (\Throwable $e) {
+            $information['requestStatus'] = new RequestStatusStruct(
+                method_exists($e, 'getErrorCode') ? $e->getErrorCode() : MigrationException::API_CONNECTION_ERROR,
+                $e->getMessage(),
+                false,
+                $e
+            );
         }
 
         return $information;
     }
 
     /**
-     * @param ReadArray $information
+     * @return array<string, mixed>
      */
-    private function doSecureCheck(array &$information): bool
+    private function getEnvironmentInformation(): array
     {
-        if ($this->client === null) {
-            return false;
-        }
-
         try {
-            $information['environmentInformation'] = $this->readData($this->client, true);
+            $data = $this->getEnvironmentData();
+        } catch (ClientException $e) {
+            if ($e->getCode() === SymfonyResponse::HTTP_NOT_FOUND) {
+                $this->checkVersion();
 
-            return true;
-        } catch (ShopwareHttpException $eVerified) {
-            $information['requestStatus'] = new RequestStatusStruct($eVerified->getErrorCode(), $eVerified->getMessage(), true);
-
-            return false;
-        }
-    }
-
-    /**
-     * @param ReadArray $information
-     */
-    private function doInsecureCheck(array &$information): bool
-    {
-        if ($this->client === null) {
-            return false;
-        }
-
-        try {
-            $information['environmentInformation'] = $this->readData($this->client);
-
-            return true;
-        } catch (ShopwareHttpException) {
-            return false;
-        }
-    }
-
-    /**
-     * @param ReadArray $information
-     */
-    private function doShopwareCheck(array &$information): bool
-    {
-        if ($this->client === null) {
-            return false;
-        }
-
-        try {
-            if ($this->checkForShopware($this->client)) {
                 throw MigrationShopwareProfileException::pluginNotInstalled();
             }
 
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
-        } catch (ShopwareHttpException $eOther) {
-            $information['requestStatus'] = new RequestStatusStruct($eOther->getErrorCode(), $eOther->getMessage());
+            throw $e;
+        }
 
-            return true;
+        return $data;
+    }
+
+    private function checkVersion(): void
+    {
+        $result = $this->doRequest('version');
+
+        $arrayResult = \json_decode($result->getBody()->getContents(), true);
+
+        if (!isset($arrayResult['success']) || $arrayResult['success'] === false) {
+            throw MigrationException::apiConnectionError('The version endpoint did not return success');
         }
     }
 
     /**
-     * @throws MigrationException
-     *
      * @return array<string, mixed>
      */
-    private function readData(HttpClientInterface $apiClient, bool $verified = false): array
+    private function getEnvironmentData(): array
     {
-        if ($verified) {
-            $apiClient = $this->connectionFactory->createApiClient($this->migrationContext, $verified);
-        }
-
-        if ($apiClient === null) {
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
-        }
-
-        $result = $this->doSecureRequest($apiClient, 'SwagMigrationEnvironment');
-
-        if ($result->getStatusCode() !== SymfonyResponse::HTTP_OK) {
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
-        }
+        $result = $this->doRequest('SwagMigrationEnvironment');
 
         $arrayResult = \json_decode($result->getBody()->getContents(), true);
 
         if (!isset($arrayResult['data'])) {
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
+            throw MigrationException::apiConnectionError('The environment endpoint did not return data');
         }
 
         return $arrayResult['data'];
     }
 
-    private function checkForShopware(HttpClientInterface $apiClient): bool
+    private function doRequest(string $endpoint): ResponseInterface
     {
-        $result = $this->doSecureRequest($apiClient, 'version');
-
-        if ($result->getStatusCode() === SymfonyResponse::HTTP_NOT_FOUND) {
-            return false;
+        if ($this->client === null) {
+            throw MigrationException::apiConnectionError(
+                'Could not create API client. Could be due to empty credentials or invalid connection.'
+            );
         }
 
-        if ($result->getStatusCode() !== SymfonyResponse::HTTP_OK) {
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
-        }
-
-        $arrayResult = \json_decode($result->getBody()->getContents(), true);
-
-        if (!isset($arrayResult['success'])
-            || (isset($arrayResult['success']) && $arrayResult['success'] === false)
-        ) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * @throws MigrationException
-     */
-    private function doSecureRequest(HttpClientInterface $apiClient, string $endpoint): ResponseInterface
-    {
         try {
-            return $apiClient->get($endpoint);
+            return $this->client->get($endpoint);
         } catch (ClientException $e) {
-            if ($e->getCode() === 401) {
+            if ($e->getCode() === SymfonyResponse::HTTP_UNAUTHORIZED) {
                 throw MigrationException::invalidConnectionAuthentication($endpoint);
             }
 
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
+            throw $e;
         } catch (GuzzleRequestException $e) {
             $response = $e->getResponse();
             if ($response !== null && \mb_strpos($response->getBody()->getContents(), 'SSL required')) {
@@ -214,9 +131,7 @@ class EnvironmentReader implements EnvironmentReaderInterface
                 throw MigrationException::requestCertificateInvalid($e->getHandlerContext()['url']);
             }
 
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
-        } catch (ConnectException $e) {
-            throw MigrationException::gatewayRead('Shopware 5.5 Api SwagMigrationEnvironment');
+            throw $e;
         }
     }
 }