+++ src/Profile/Shopware6/Gateway/Api/Reader/EnvironmentReader.php
@@ -10,7 +10,6 @@ namespace SwagMigrationAssistant\Profile\Shopware6\Gateway\Api\Reader;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
 use Shopware\Core\Framework\Log\Package;
-use Shopware\Core\Framework\ShopwareHttpException;
 use SwagMigrationAssistant\Exception\MigrationException;
 use SwagMigrationAssistant\Migration\Gateway\HttpClientInterface;
 use SwagMigrationAssistant\Migration\Gateway\Reader\EnvironmentReaderInterface;
@@ -25,7 +24,7 @@ use Symfony\Component\HttpFoundation\Response as SymfonyResponse;
 #[Package('services-settings')]
 class EnvironmentReader implements EnvironmentReaderInterface
 {
-    private ?HttpClientInterface $client;
+    private ?HttpClientInterface $client = null;
 
     public function __construct(private readonly ConnectionFactoryInterface $connectionFactory)
     {
@@ -43,82 +42,36 @@ class EnvironmentReader implements EnvironmentReaderInterface
             'requestStatus' => new RequestStatusStruct(),
         ];
 
-        if ($this->client === null) {
-            $information['requestStatus'] = new RequestStatusStruct('SWAG-EMPTY-CREDENTIALS', 'Empty credentials');
-
-            return $information;
-        }
-
-        if ($this->doSecureCheck($information)) {
-            return $information;
-        }
-
-        $requestStatus = $information['requestStatus'];
-
-        if ($requestStatus->getCode() === MigrationException::sslRequired()->getErrorCode()) {
-            $requestStatus->setIsWarning(false);
-
-            return $information;
-        }
-
-        if ($this->doInsecureCheck($information)) {
-            return $information;
+        try {
+            $information['environmentInformation'] = $this->getEnvironment();
+        } catch (\Throwable $e) {
+            $information['requestStatus'] = new RequestStatusStruct(
+                method_exists($e, 'getErrorCode') ? $e->getErrorCode() : MigrationException::API_CONNECTION_ERROR,
+                $e->getMessage(),
+                false,
+                $e
+            );
         }
 
         return $information;
     }
 
     /**
-     * @param ReadArray $information
-     */
-    private function doSecureCheck(array &$information): bool
-    {
-        try {
-            $information['environmentInformation'] = $this->readData(true);
-
-            return true;
-        } catch (ShopwareHttpException $eVerified) {
-            $information['requestStatus'] = new RequestStatusStruct($eVerified->getErrorCode(), $eVerified->getMessage(), false);
-
-            return false;
-        }
-    }
-
-    /**
-     * @param ReadArray $information
-     */
-    private function doInsecureCheck(array &$information): bool
-    {
-        try {
-            $information['environmentInformation'] = $this->readData();
-
-            return true;
-        } catch (ShopwareHttpException $eUnverified) {
-            $information['requestStatus'] = new RequestStatusStruct($eUnverified->getErrorCode(), $eUnverified->getMessage(), false);
-
-            return false;
-        }
-    }
-
-    /**
      * @return array<string, mixed>
      */
-    private function readData(bool $verified = false): array
+    private function getEnvironment(): array
     {
         if ($this->client === null) {
-            return [];
+            throw MigrationException::apiConnectionError(
+                'Could not create API client. Could be due to empty credentials or invalid connection.'
+            );
         }
 
         try {
-            $result = $this->client->get(
-                'get-environment',
-                [
-                    'verify' => $verified,
-                ]
-            );
+            $result = $this->client->get('get-environment');
 
             if ($result->getStatusCode() !== SymfonyResponse::HTTP_OK) {
-                throw MigrationException::gatewayRead('Shopware 6 API Environment Call');
+                throw MigrationException::apiConnectionError('The response status code was not 200.');
             }
 
             return \json_decode($result->getBody()->getContents(), true);
@@ -127,7 +80,7 @@ class EnvironmentReader implements EnvironmentReaderInterface
                 throw MigrationException::invalidConnectionAuthentication('get-data');
             }
 
-            throw MigrationException::gatewayRead('Shopware 6 API Environment Call');
+            throw $e;
         } catch (RequestException $e) {
             if ($e->getRequest()->getUri()->getPath() === '/api/oauth/token') {
                 // something went wrong with authentication.
@@ -143,7 +96,7 @@ class EnvironmentReader implements EnvironmentReaderInterface
                 throw MigrationException::requestCertificateInvalid($e->getHandlerContext()['url']);
             }
 
-            throw MigrationException::gatewayRead('Shopware 6 API Environment Call');
+            throw $e;
         }
     }
 }